+++ includes/_options.php
@@ -41,6 +41,7 @@ function eduadmin_settings_init()
 	register_setting('eduadmin-booking', 'eduadmin-allowInterestRegEvent');
 	register_setting('eduadmin-booking', 'eduadmin-hideSubEventDateTime');
 	register_setting('eduadmin-booking', 'eduadmin-allowDiscountCode');
+	register_setting('eduadmin-booking', 'eduadmin-noInvoiceFreeEvents');
 
 	/* Phrase settings */
 	register_setting('eduadmin-phrases', 'eduadmin-phrases');